refactor: 확장함수를 별도의 파일에서 관리
