+++ app/src/main/java/woowacourse/shopping/ui/payment/CouponUiModel.kt
@@ -6,5 +6,3 @@ data class CouponUiModel(
     val couponState: CouponState,
     val isChecked: Boolean = false,
 )
-
-fun CouponState.toUiModel() = CouponUiModel(couponState = this)
