+++ app/src/main/java/woowacourse/shopping/ui/payment/Extension.kt
@@ -1,0 +1,5 @@
+package woowacourse.shopping.ui.payment
+
+import woowacourse.shopping.data.coupon.CouponState
+
+fun CouponState.toUiModel() = CouponUiModel(couponState = this)
