+++ app/src/main/java/woowacourse/shopping/ui/payment/viewmodel/PaymentViewModel.kt
@@ -105,8 +105,7 @@ class PaymentViewModel(
         }
     }
 
-    private fun getCoupons() =
-        requireNotNull(_coupons.value?.map { it.copy(isChecked = false) }?.toMutableList())
+    private fun getCoupons() = requireNotNull(_coupons.value?.map { it.copy(isChecked = false) }?.toMutableList())
 
     private fun findCheckedCoupon(couponId: Long): CouponUiModel {
         return requireNotNull(_coupons.value?.find { it.couponState.coupon.id == couponId })
