+++ app/src/main/java/woowacourse/shopping/ui/products/Extension.kt
@@ -1,0 +1,3 @@
+package woowacourse.shopping.ui.products
+
+fun List<ProductUiModel>.isLoading() = this + LoadingUiModel + LoadingUiModel
